Renamed pagesto index.js as per code review
