+++ src/App.js
@@ -1,24 +1,24 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
-import Login from './Pages/LoginPage/Login';
+import Login from './Pages/LoginPage';
 import SuccessAnimation from "./components/SuccessAnimation/SuccessAnimation";
 import TOS from "./components/TosAndPrivacyPolicy/TOS";
-import Dashboard from './Pages/Dashboard/Dashboard';
+import Dashboard from './Pages/Dashboard';
 import Homepage from './components/Homepage/Homepage';
 import Category from './components/Category/Category';
 import CategoryDetails from './components/CategoryDetails/CategoryDetails';
 import Product from './components/Product/Product';
-import CreateListing from './Pages/CreateListing/CreateListing';
+import CreateListing from './Pages/CreateListing';
 import CreateListingCard from './components/CreateListingCard/CreateListingCard';
 import UploadImageModal from './components/UploadImageModal/UploadImageModal';
-import Account from './Pages/AccountInformation/Account';
-import AccountInfo from './components/AccountInfo/AccountInfo';
+import Account from './Pages/AccountInformation';
+import AccountInfo from './components/AccountInfo';
 import ChangePassword from './components/ChangePassword/ChangePassword';
-import MyListings from './Pages/MyListings/MyLisings';
+import MyListings from './Pages/MyListings';
 import MyListingDetails from './components/MyListingDetails/MyListingDetails';
 import Dummydata from "./dummyData.json";
 import MyListingEdit from './components/MyListingEdit/MyListingEdit';
-import MyPurchase from './Pages/MyPurchase/MyPurchase';
+import MyPurchase from './Pages/MyPurchase';
 import MyPurchases from './components/MyPurchases/MyPurchases';
 import PickUpInformation from './components/PickUpInformation/PickUpInformation';
 import ProductDetails from './components/ProductDetails/ProductDetails';
